Clean up all references to Evergreen costs
