--- cost/testdata/spend_test.yml
@@ -1,26 +1,0 @@
-options:
-  duration: 4h
-  directory: testdata
-evergreen:
-  user: user
-  key: key
-  url: http://rooturl.com
-aws:
-  s3:
-    s3key_start: brian.samek/s3_testdata
-    s3bucket: boxes.10gen.com/
-  accounts:
-    - name: kernel-build
-      key: "foo-key"
-      secret: "foo-bar"
-  ebs_pricing:
-    gp2: 0.10
-    io1GB: 0.125
-    io1IOPS: 0.065
-    st1: 0.045
-    sc1: 0.025
-    standard: 0.05
-    snapshot: 0.05
-providers:
-  - name: fakecompany
-    cost: 50000.0
