--- cost/testdata/spot-price-data.txt
@@ -1,67 +1,0 @@
-{
-    "SpotPriceHistory": [
-        {
-            "Timestamp": "2017-07-19T09:15:46.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043900"
-        },
-        {
-            "Timestamp": "2017-07-19T09:14:01.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043800"
-        },
-        {
-            "Timestamp": "2017-07-19T07:16:33.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043900"
-        },
-        {
-            "Timestamp": "2017-07-19T07:14:03.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043800"
-        },
-        {
-            "Timestamp": "2017-07-19T06:53:25.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043900"
-        },
-        {
-            "Timestamp": "2017-07-19T06:50:02.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043800"
-        },
-        {
-            "Timestamp": "2017-07-19T06:22:01.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043900"
-        },
-        {
-            "Timestamp": "2017-07-19T06:19:02.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.044000"
-        },
-        {
-            "Timestamp": "2017-07-19T06:11:35.000Z",
-            "AvailabilityZone": "us-east-1a",
-            "InstanceType": "c3.xlarge",
-            "ProductDescription": "Linux/UNIX",
-            "SpotPrice": "0.043900"
-        }
-    ]
-}
