+++ evergreen/doc.go
@@ -1,0 +1,4 @@
+/*
+Package evergreen provides helpers to interact with the Evergreen API.
+*/
+package evergreen
