+++ model/depgraph_edge_test.go
@@ -23,7 +23,7 @@ func TestGraphEdge(t *testing.T) {
 	}()
 
 	for name, test := range map[string]func(context.Context, *testing.T, cedar.Environment, *GraphEdge){
-		// "": func(ctx context.Context, t *testing.t, env cedar.Environment, conf *Costreport) {},
+		// "": func(ctx context.Context, t *testing.t, env cedar.Environment, e *GraphEdge) {},
 	} {
 		t.Run(name, func(t *testing.T) {
 			tctx, cancel := context.WithCancel(ctx)
