+++ model/sender_test.go
@@ -23,7 +23,7 @@ func TestEvent(t *testing.T) {
 	}()
 
 	for name, test := range map[string]func(context.Context, *testing.T, cedar.Environment, *Event){
-		// "": func(ctx context.Context, t *testing.t, env cedar.Environment, conf *Costreport) {},
+		// "": func(ctx context.Context, t *testing.t, env cedar.Environment, e *Event) {},
 	} {
 		t.Run(name, func(t *testing.T) {
 			tctx, cancel := context.WithCancel(ctx)
